certificate verification using SHA-512/requirements.txt
Flask==2.2.3
Flask-MySQLdb==1.0.1
pycryptodome==3.17 
fpdf==1.7.2             
qrcode==7.3.1           
Pillow==9.5.0           